Created some abstractings (adding ssl support in future)
